make pyjs a lib again
