+++ src/export_pyjs_module.cpp
@@ -1,5 +1,3 @@
-#pragma once
-
 #include <pybind11/pybind11.h>
 
 #include <emscripten.h>
